Minor fixes
Cleaning up here and there.
