+++ .editorconfig
@@ -1,0 +1,26 @@
+# SPDX-FileCopyrightText: 2008-2023 HPDCS Group <rootsim@googlegroups.com>
+# SPDX-License-Identifier: CC0-1.0
+
+root = true
+
+[*]
+trim_trailing_whitespace = true
+charset = utf-8
+end_of_line = lf
+insert_final_newline = true
+
+[*.[ch]]
+indent_style = tab
+indent_size = 8
+
+[*.cpp]
+indent_style = tab
+indent_size = 8
+
+[Makefile]
+indent_style = tab
+indent_size = 8
+
+[*.py]
+indent_style = space
+indent_size = 4
