+++ pseudosql/src/test_root_sim/check.py
